Migrate invite_1_claimer tests
