+++ libparsec/crates/protocol/tests/invite.rs
@@ -427,92 +427,6 @@ fn serde_invite_info_rep(#[case] raw: &[u8], #[case] expected: invited_cmds::inv
 }
 
 #[parsec_test]
-fn serde_invite_1_claimer_wait_peer_req() {
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   claimer_public_key: hex!("6507907d33bae6b5980b32fa03f3ebac56141b126e44f352ea46c5f22cd5ac57")
-    //   cmd: "invite_1_claimer_wait_peer"
-    let raw = hex!(
-        "82b2636c61696d65725f7075626c69635f6b6579c4206507907d33bae6b5980b32fa03f3eb"
-        "ac56141b126e44f352ea46c5f22cd5ac57a3636d64ba696e766974655f315f636c61696d65"
-        "725f776169745f70656572"
-    );
-
-    let req = invited_cmds::invite_1_claimer_wait_peer::Req {
-        claimer_public_key: PublicKey::from(hex!(
-            "6507907d33bae6b5980b32fa03f3ebac56141b126e44f352ea46c5f22cd5ac57"
-        )),
-    };
-
-    let expected = invited_cmds::AnyCmdReq::Invite1ClaimerWaitPeer(req);
-
-    let data = invited_cmds::AnyCmdReq::load(&raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = if let invited_cmds::AnyCmdReq::Invite1ClaimerWaitPeer(data) = data {
-        data.dump().unwrap()
-    } else {
-        unreachable!()
-    };
-
-    let data2 = invited_cmds::AnyCmdReq::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
-#[case::ok(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   greeter_public_key: hex!("6507907d33bae6b5980b32fa03f3ebac56141b126e44f352ea46c5f22cd5ac57")
-    //   status: "ok"
-    &hex!(
-        "82b2677265657465725f7075626c69635f6b6579c4206507907d33bae6b5980b32fa03f3eb"
-        "ac56141b126e44f352ea46c5f22cd5ac57a6737461747573a26f6b"
-    )[..],
-    invited_cmds::invite_1_claimer_wait_peer::Rep::Ok {
-        greeter_public_key: PublicKey::from(hex!(
-            "6507907d33bae6b5980b32fa03f3ebac56141b126e44f352ea46c5f22cd5ac57"
-        )),
-    }
-)]
-#[case::not_found(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "not_found"
-    &hex!(
-        "81a6737461747573a96e6f745f666f756e64"
-    )[..],
-    invited_cmds::invite_1_claimer_wait_peer::Rep::NotFound
-)]
-#[case::invalid_state(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "invalid_state"
-    &hex!(
-        "81a6737461747573ad696e76616c69645f7374617465"
-    )[..],
-    invited_cmds::invite_1_claimer_wait_peer::Rep::InvalidState
-)]
-fn serde_invite_1_claimer_wait_peer_rep(
-    #[case] raw: &[u8],
-    #[case] expected: invited_cmds::invite_1_claimer_wait_peer::Rep,
-) {
-    let data = invited_cmds::invite_1_claimer_wait_peer::Rep::load(raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = data.dump().unwrap();
-
-    let data2 = invited_cmds::invite_1_claimer_wait_peer::Rep::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
 fn serde_invite_1_greeter_wait_peer_req() {
     // Generated from Python implementation (Parsec v2.6.0+dev)
     // Content:
